Fix ClassNotFoundException

+++ src/main/java/io/github/zekerzhayard/skinmodcompact/asm/mixins/misc/ThreadCleanCache.java
@@ -3,6 +3,8 @@ package io.github.zekerzhayard.skinmodcompact.asm.mixins.misc;
 import customskinloader.CustomSkinLoader;
 import customskinloader.utils.HttpRequestUtil;
 import customskinloader.utils.HttpTextureUtil;
+import net.minecraft.launchwrapper.Launch;
+import net.minecraft.launchwrapper.LaunchClassLoader;
 import org.apache.commons.io.FileUtils;
 
 public class ThreadCleanCache extends Thread {
@@ -12,6 +14,15 @@ public class ThreadCleanCache extends Thread {
 
     @Override
     public void run() {
+        if (!(this.getClass().getClassLoader() instanceof LaunchClassLoader)) {
+            try {
+                Class<?> cl = Launch.classLoader.loadClass(this.getClass().getName());
+                cl.getMethod("run").invoke(cl.newInstance());
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            return;
+        }
         try {
             FileUtils.deleteDirectory(HttpRequestUtil.CACHE_DIR);
             FileUtils.deleteDirectory(HttpTextureUtil.getCacheDir());
